+++ src/pre_derived_keys_cache/mod.rs
@@ -1,4 +1,7 @@
 #[allow(clippy::module_inception)]
 mod pre_derived_keys_cache;
 
+mod cache_outcome;
+
+pub use cache_outcome::*;
 pub use pre_derived_keys_cache::*;
