+++ src/recovery_securify_cache/helpers/mod.rs
@@ -1,12 +1,18 @@
+mod any_factor_derivation_request;
+mod any_factor_derivation_requests;
 mod derivation_analysis_types;
 mod derivation_request;
+mod derivation_requests;
 mod factor_instances;
 mod is_derivation_done_query;
 mod onchain_analyzer;
 mod profile_analyzer;
 
+pub use any_factor_derivation_request::*;
+pub use any_factor_derivation_requests::*;
 pub use derivation_analysis_types::*;
 pub use derivation_request::*;
+pub use derivation_requests::*;
 pub use factor_instances::*;
 pub use is_derivation_done_query::*;
 pub use onchain_analyzer::*;
